A log of changes related to Artifactory and tools
Role 'artifactory' was implemented but is NOT complete!
I've switched to use Ansible collection jfrog.platform
that is managed by JFrog. It was a big pain to use it but
finally I was able to run artifactory together with NGINX
and PostgreSQL. So this role should NOT be used

As the roles by jfrog.platform did not apply working setup
I did it manually. To have the working configuration I have
created additional role 'artifactory-conf' that keeps the
working configuration. NOTE that this role is never tested!

Added tools-linux role that setups linux machine to be used
by starting the jobs instead of using the Jenkins controller.
This is some sort of security - to isolate the controller so
it cannot run anything.

Minor changes to build-win role related to Octopus and JFrog

Added README files where applicable