Added UploadFiles.ps1
The script was implemented primarly to upload NuGets from
the network drive \\bg-pld-nas001\ReconArt.Nuget to the new
artifactory:
https://artifactory.reconart.net:443/artifactory/generic-reconart-nuget/

It could be used to sync another directories but need to use
propper values for the parameters.

By now this script can be run only from a machine that is in the
domain (to have access to \\bg-pld\nas001) and the user needs the
password for user 'jenkins'. It is in ansible-vault
(ansible/vars/artifactory-secret-vars.yaml)